+++ __converter/app/calculate-eta/README.md
@@ -1,0 +1,6 @@
+# Calculate ETA
+
+TODO: Add explanation about what the file does
+
+This file is not called directly because node does not support all the used es6 features.
+The file `calculateETA.js` in the parent directory contains the babel register and should be executed instead.
